requirements.txt
streamlit==1.38.0
requests>=2.32
pandas>=2.2
numpy>=1.26
pydantic>=2.7
pyportfolioopt>=1.5.6
scikit-learn>=1.5
joblib>=1.3
python-dotenv>=1.0
SQLAlchemy>=2.0
pre-commit>=3.7
ruff>=0.6
pytest>=8.3
